+++ app/src/main/java/de/tap/easy_xkcd/fragments/ComicBrowserFragment.java
@@ -45,6 +45,7 @@ import java.io.IOException;
 import java.util.Arrays;
 
 import de.tap.easy_xkcd.Activities.MainActivity;
+import de.tap.easy_xkcd.database.DatabaseManager;
 import de.tap.easy_xkcd.database.RealmComic;
 import de.tap.easy_xkcd.utils.Comic;
 import de.tap.easy_xkcd.utils.JsonParser;
@@ -311,7 +312,7 @@ public class ComicBrowserFragment extends ComicFragment {
             case R.id.action_random:
                 return getRandomComic();
             case R.id.action_thread:
-                return showThread(lastComicNumber);
+                return DatabaseManager.showThread(comicMap.get(lastComicNumber).getComicData()[0], getActivity(), false);
         }
         return super.onOptionsItemSelected(item);
     }